+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.4.3-5] - 2020-01-12
+### Added
+- set outline in world
+### Fixed
+- output outline maintile
+
 ## [0.4.3-4] - 2020-01-12
 ### Added
 - note and titles counter
@@ -250,7 +256,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Example story as usage.
 - Output story as markdown.
 
-[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.4.3-4...HEAD
+[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.4.3-5...HEAD
+[0.4.3-5]: https://github.com/nagisc007/storybuilder/releases/v0.4.3-5
 [0.4.3-4]: https://github.com/nagisc007/storybuilder/releases/v0.4.3-4
 [0.4.3-3]: https://github.com/nagisc007/storybuilder/releases/v0.4.3-3
 [0.4.3-2]: https://github.com/nagisc007/storybuilder/releases/v0.4.3-2